backend/requirements.txt
ultralytics
opencv-python
torch
torchvision
timm==0.6.13
fastapi
uvicorn[standard]
PyTurboJPEG
Pillow
google-genai
pydantic>=2
